+++ meridian/analysis/optimizer.py
@@ -1278,7 +1278,8 @@ class BudgetOptimizer:
       self,
       new_data: analyzer.DataTensors | None = None,
       use_posterior: bool = True,
-      selected_times: tuple[str | None, str | None] | None = None,
+      start_date: tc.Date = None,
+      end_date: tc.Date = None,
       fixed_budget: bool = True,
       budget: float | None = None,
       pct_of_spend: Sequence[float] | None = None,
@@ -1332,15 +1333,16 @@ class BudgetOptimizer:
       use_posterior: Boolean. If `True`, then the budget is optimized based on
         the posterior distribution of the model. Otherwise, the prior
         distribution is used.
-      selected_times: Tuple containing the start and end time dimension
-        coordinates for the duration to run the optimization on. Selected time
-        values should align with the Meridian time dimension coordinates in the
-        underlying model if optimizing the original data. If `new_data` is
-        provided with a different number of time periods than in `InputData`,
-        then the start and end time coordinates must match the time dimensions
-        in `new_data.time`. By default, all times periods are used. Either start
-        or end time component can be `None` to represent the first or the last
-        time coordinate, respectively.
+      start_date: Optional start date selector, *inclusive*, in _yyyy-mm-dd_
+        format. Default is `None`, i.e. the first time period.
+      end_date: Optional end date selector, *inclusive* in _yyyy-mm-dd_ format.
+        Both start and end time selectors should align with the Meridian time
+        dimension coordinates in the underlying model if optimizing the original
+        data. If `new_data` is provided with a different number of time periods
+        than in `InputData`, then the start and end time coordinates must match
+        the time dimensions in `new_data.time`. By default, all times periods
+        are used. Either start or end time component can be `None` to represent
+        the first or the last time coordinate, respectively.
       fixed_budget: Boolean indicating whether it's a fixed budget optimization
         or flexible budget optimization. Defaults to `True`. If `False`, must
         specify either `target_roi` or `target_mroi`.
@@ -1420,7 +1422,8 @@ class BudgetOptimizer:
     use_grid_arg = optimization_grid is not None and self._validate_grid(
         new_data=new_data,
         use_posterior=use_posterior,
-        selected_times=selected_times,
+        start_date=start_date,
+        end_date=end_date,
         budget=budget,
         pct_of_spend=pct_of_spend,
         spend_constraint_lower=spend_constraint_lower,
@@ -1433,7 +1436,8 @@ class BudgetOptimizer:
     if optimization_grid is None or not use_grid_arg:
       optimization_grid = self.create_optimization_grid(
           new_data=new_data,
-          selected_times=selected_times,
+          start_date=start_date,
+          end_date=end_date,
           budget=budget,
           pct_of_spend=pct_of_spend,
           spend_constraint_lower=spend_constraint_lower,
@@ -1545,7 +1549,8 @@ class BudgetOptimizer:
       self,
       new_data: analyzer.DataTensors | None,
       use_posterior: bool,
-      selected_times: tuple[str | None, str | None] | None,
+      start_date: tc.Date,
+      end_date: tc.Date,
       budget: float | None,
       pct_of_spend: Sequence[float] | None,
       spend_constraint_lower: _SpendConstraint,
@@ -1584,15 +1589,16 @@ class BudgetOptimizer:
       )
       return False
 
-    selected_time_dims = self._validate_selected_times(
-        selected_times=selected_times,
+    selected_times = self._validate_selected_times(
+        start_date=start_date,
+        end_date=end_date,
         new_data=new_data,
     )
-    if not np.array_equal(selected_time_dims, optimization_grid.selected_times):
+    if not np.array_equal(selected_times, optimization_grid.selected_times):
       warnings.warn(
           'Given optimization grid was created with `selected_times` ='
           f' {optimization_grid.selected_times}, but optimization was called'
-          f' with `selected_times` = {selected_time_dims}. A new grid will be'
+          f' with `selected_times` = {selected_times}. A new grid will be'
           ' created.'
       )
       return False
@@ -1614,9 +1620,14 @@ class BudgetOptimizer:
       return False
 
     n_channels = len(optimization_grid.channels)
+    selected_times = self._validate_selected_times(
+        start_date=start_date,
+        end_date=end_date,
+        new_data=new_data,
+    )
     hist_spend = self._analyzer.get_aggregated_spend(
         new_data=filled_data.filter_fields(c.PAID_CHANNELS + c.SPEND_DATA),
-        selected_times=selected_time_dims,
+        selected_times=selected_times,
         include_media=self._meridian.n_media_channels > 0,
         include_rf=self._meridian.n_rf_channels > 0,
     ).data
@@ -1664,7 +1675,8 @@ class BudgetOptimizer:
       self,
       new_data: xr.Dataset | None = None,
       use_posterior: bool = True,
-      selected_times: tuple[str | None, str | None] | None = None,
+      start_date: tc.Date = None,
+      end_date: tc.Date = None,
       budget: float | None = None,
       pct_of_spend: Sequence[float] | None = None,
       spend_constraint_lower: _SpendConstraint = c.SPEND_CONSTRAINT_DEFAULT,
@@ -1689,15 +1701,16 @@ class BudgetOptimizer:
       use_posterior: Boolean. If `True`, then the incremental outcome is derived
         from the posterior distribution of the model. Otherwise, the prior
         distribution is used.
-      selected_times: Tuple containing the start and end time dimension
-        coordinates for the duration to run the optimization on. Selected time
-        values should align with the Meridian time dimension coordinates in the
-        underlying model if optimizing the original data. If `new_data` is
-        provided with a different number of time periods than in `InputData`,
-        then the start and end time coordinates must match the time dimensions
-        in `new_data.time`. By default, all times periods are used. Either start
-        or end time component can be `None` to represent the first or the last
-        time coordinate, respectively.
+      start_date: Optional start date selector, *inclusive*, in _yyyy-mm-dd_
+        format. Default is `None`, i.e. the first time period.
+      end_date: Optional end date selector, *inclusive* in _yyyy-mm-dd_ format.
+        Both start and end time selectors should align with the Meridian time
+        dimension coordinates in the underlying model if optimizing the original
+        data. If `new_data` is provided with a different number of time periods
+        than in `InputData`, then the start and end time coordinates must match
+        the time dimensions in `new_data.time`. By default, all times periods
+        are used. Either start or end time component can be `None` to represent
+        the first or the last time coordinate, respectively.
       budget: Number indicating the total budget for the fixed budget scenario.
         Defaults to the historical budget.
       pct_of_spend: Numeric list of size `n_paid_channels` containing the
@@ -1751,12 +1764,14 @@ class BudgetOptimizer:
     filled_data = new_data.validate_and_fill_missing_data(
         required_tensors_names=required_tensors, meridian=self._meridian
     )
-    selected_time_dims = self._validate_selected_times(
-        selected_times, filled_data
+    selected_times = self._validate_selected_times(
+        start_date=start_date,
+        end_date=end_date,
+        new_data=filled_data,
     )
     hist_spend = self._analyzer.get_aggregated_spend(
         new_data=filled_data.filter_fields(c.PAID_CHANNELS + c.SPEND_DATA),
-        selected_times=selected_time_dims,
+        selected_times=selected_times,
         include_media=self._meridian.n_media_channels > 0,
         include_rf=self._meridian.n_rf_channels > 0,
     ).data
@@ -1783,7 +1798,7 @@ class BudgetOptimizer:
           self._analyzer.optimal_freq(
               new_data=filled_data.filter_fields(c.RF_DATA),
               use_posterior=use_posterior,
-              selected_times=selected_time_dims,
+              selected_times=selected_times,
               use_kpi=use_kpi,
           ).optimal_frequency,
           dtype=tf.float32,
@@ -1797,7 +1812,7 @@ class BudgetOptimizer:
         spend_bound_lower=optimization_lower_bound,
         spend_bound_upper=optimization_upper_bound,
         step_size=step_size,
-        selected_times=selected_time_dims,
+        selected_times=selected_times,
         new_data=filled_data.filter_fields(c.PAID_DATA),
         use_posterior=use_posterior,
         use_kpi=use_kpi,
@@ -1819,7 +1834,7 @@ class BudgetOptimizer:
         gtol=gtol,
         round_factor=round_factor,
         optimal_frequency=optimal_frequency,
-        selected_times=selected_time_dims,
+        selected_times=selected_times,
     )
 
   def _create_grid_dataset(
@@ -1864,13 +1879,11 @@ class BudgetOptimizer:
 
   def _validate_selected_times(
       self,
-      selected_times: tuple[str | None, str | None] | None,
+      start_date: tc.Date,
+      end_date: tc.Date,
       new_data: analyzer.DataTensors | None,
   ) -> Sequence[str] | Sequence[bool] | None:
     """Validates and returns the selected times."""
-    if selected_times is None:
-      return None
-    start_date, end_date = selected_times
     if start_date is None and end_date is None:
       return None